+++ packet-codec/src/main/java/io/bedrockbridge/protocol/codec/PacketCodec.java
@@ -1,0 +1,6 @@
+package io.bedrockbridge.protocol.codec;
+
+import io.bedrockbridge.protocol.Packet;
+
+/** Combined encoder and decoder for one packet class. */
+public interface PacketCodec<T extends Packet> extends PacketEncoder<T>, PacketDecoder<T> {}